+++ test/view.js
@@ -109,7 +109,7 @@ test("remove child", function () {
 });
 
 
-test("bind with data", function () {
+test("bind with model", function () {
     var view, model;
 
     model = new yamvc.Model({
@@ -195,4 +195,53 @@ test("rendered two times", function () {
     model.data('display', 'block');
 
     equal(view.queryEl('.example').getAttribute('style'), 'display: block;');
+});
+
+test("bind with collection", function () {
+    var view, collection;
+
+    collection = new yamvc.Collection({
+        config: {
+            namespace: 'example',
+            data: [
+                {
+                    name: "Seba",
+                    display: "true"
+                },
+                {
+                    name: "Seba 2",
+                    display: "none"
+                },
+                {
+                    name: "Seba 3",
+                    display: "true"
+                }
+            ]
+        }
+    });
+
+    view = new yamvc.View({
+        config: {
+            collections: [
+                collection
+            ],
+            tpl: new yamvc.view.Template({
+                config: {
+                    id: 'tpl-example-5',
+                    tpl: [
+                        '<div>',
+                        '<ul reapeat="">',
+                        '<li>{{}}</li>',
+                        '</ul>',
+                        '</div>'
+                    ]
+                }
+            }),
+            renderTo: '#test-6'
+        }
+    });
+
+    view.render();
+
+    equal(view.queryEl('.example').innerText, 'Hi Seba');
 });